backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
paramiko==3.4.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
